requirements.txt
fastapi==0.89.1
fastapi-security==0.5.0
uvicorn==0.20.0
asyncpg==0.27.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
alembic==1.8.1
python-dateutil==2.8.2
orjson==3.8.5

sqlmodel==0.0.8
aioredis==2.0.1
# fastapi-admin==1.0.4
# fastapi_amis_admin==0.4.2
sqladmin==0.8.0
email-validator==1.3.0
# Jinja2==3.1.2
# pydantic==1.10.2

httpx==0.23.1
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-async-sqlalchemy==0.2.0
ipython==8.7.0
fastapi-debug-toolbar==0.3.2
freezegun==1.2.2

isort==5.11.4
black==22.12.0
mypy==0.991